Updates infra to latest CDK
Using the prior project as a reference, this commit updates
the AWS CDK dependency to the latest version and makes the
relevant changes (thankfully mostly simplifications) to support
latest CDK.

In particular, this includes leveraging the crossRegionReferences
property of Stacks to replace the old manual approach of creating
SSM Custom Resources to pass values (e.g. ARNs) between regions.

This commit also removes the lambda@edge functionality entirely,
as it wasn't yet being used for anything. This may be added back
later (e.g. to support OIDC auth), but for now it's not necessary.